Move integration tests to the appropriate test file
