+++ src/test/java/seedu/address/logic/commands/AddJournalEntryCommandTest.java
@@ -4,11 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_AMY;
 import static seedu.address.testutil.Assert.assertThrows;
-import static seedu.address.testutil.TypicalEntries.getTypicalJournal;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -25,19 +21,14 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.ValidCommand;
 import seedu.address.model.AddressBook;
-import seedu.address.model.AliasMap;
-import seedu.address.model.Journal;
 import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyAliasMap;
 import seedu.address.model.ReadOnlyJournal;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.UserPrefs;
 import seedu.address.model.journal.Entry;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.EntryBuilder;
-import seedu.address.testutil.PersonBuilder;
 
 class AddJournalEntryCommandTest {
     /**
@@ -255,33 +246,6 @@ class AddJournalEntryCommandTest {
             assertEquals(Arrays.asList(validEntry), modelStub.entriesAdded);
         }
 
-        @Test
-        @DisplayName("should add entry with contacts successfully if contacts"
-                + " is in model")
-        public void execute_entryWithContactsAcceptedByModel_addSuccessful()
-                throws CommandException{
-            Model model = new ModelManager(
-                    getTypicalAddressBook(),
-                    new Journal(),
-                    new UserPrefs(),
-                    new AliasMap()
-            );
-            // ALICE is in typicalAddressBook
-            Person toAdd = ALICE;
-            Entry validEntry = new EntryBuilder()
-                    .withContacts(toAdd)
-                    .build();
-            AddJournalEntryCommand addCommand =
-                    new AddJournalEntryCommand(validEntry);
-            CommandResult commandResult =
-                    addCommand.execute(model);
-            assertEquals(
-                    String.format(AddJournalEntryCommand.MESSAGE_SUCCESS,
-                            validEntry),
-                    commandResult.getFeedbackToUser()
-            );
-            assertEquals(Arrays.asList(validEntry), model.getJournal().getEntryList());
-        }
 
         @Test
         @DisplayName("should throw CommandException if entry is already in "
@@ -297,35 +261,6 @@ class AddJournalEntryCommandTest {
                     AddJournalEntryCommand.MESSAGE_DUPLICATE_ENTRY, () ->
                             addCommand.execute(modelStub)
             );
-        }
-
-        @Test
-        @DisplayName("should throw CommandException if contact specified is "
-                + "not in model")
-        public void execute_contactNotInModel_throwsCommandException() {
-            // default PersonBuilder is Alice Pauline
-            Person toAdd = new PersonBuilder()
-                    .withName(VALID_NAME_AMY)
-                    .build();
-            Entry validEntry = new EntryBuilder()
-                    .withContacts(toAdd)
-                    .build();
-            Model model = new ModelManager(
-                    getTypicalAddressBook(),
-                    getTypicalJournal(),
-                    new UserPrefs(),
-                    new AliasMap()
-            );
-            AddJournalEntryCommand addCommand = new AddJournalEntryCommand(validEntry);
-            String expectedMessage =
-                    String.format(AddJournalEntryCommand.MESSAGE_PERSON_NOT_FOUND,
-                            toAdd.getName());
-
-            assertThrows(CommandException.class,
-                    expectedMessage,
-                    () -> addCommand.execute(model)
-            );
-
         }
     }
